+++ include/GafferScene/Deformer.h
@@ -80,6 +80,26 @@ class GAFFERSCENE_API Deformer : public ObjectProcessor
 		/// > accessed by `adjustBounds()`.
 		virtual bool adjustBounds() const;
 
+		/// If `computeProcessedObjectBound()` is overridden, this must be overriden
+		/// to return true for any plugs it uses. Unlike other affects methods, overrides
+		/// should _not_ call the base class implementation.
+		virtual bool affectsProcessedObjectBound( const Gaffer::Plug *input ) const;
+		/// If `computeProcessedObjectBound()` is overridden, this must be
+		/// be overridden to match. Unlike other hash methods, overrides should
+		/// _not_ call the base class implementation.
+		virtual void hashProcessedObjectBound( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const;
+		/// May be implemented by derived classes to return a bound for the
+		/// result of `computeProcessedObject()`. This will only be called if
+		/// `adjustBounds()` returns true. The default implementation uses
+		/// the brute force approach of actually processing the object, so
+		/// reimplementing to provide a cheaper approximate bound may improve
+		/// performance considerably.
+		/// > Note : Implementations are currently hampered by the fact that
+		/// > `in.bound` provides the bound for the input object _and_ its
+		/// > children. We could consider having separate `in.objectBound`
+		/// > and `in.childBound` plugs instead.
+		virtual Imath::Box3f computeProcessedObjectBound( const ScenePath &path, const Gaffer::Context *context ) const;
+
 	private :
 
 		void init();